tools/categorizer/requirements.txt
scikit-learn>=1.4,<2.0
skl2onnx>=1.16,<2.0
onnxruntime>=1.17,<2.0
numpy>=1.24,<3.0
pandas>=2.0,<3.0
supabase>=2.0,<3.0
python-dotenv>=1.0,<2.0
